Instead of activating the conda env, just call interpreter directly
Slurm appears to have a bug where running an array of jobs that all
call source activate <env> sometimes fails, with the result that all
of the jobs fail because they are unable to load required packages.

Here, I attempt to invoke the Python interpreter for the eclips_venv
environment directly to see if it fixes the problem.